+++ templates/index.php
@@ -24,7 +24,7 @@
                     </div>
                     <div class="lot__info">
                         <span class="lot__category"><?=strip_tags($lot["category"]);?></span>
-                        <h3 class="lot__title"><a class="text-link" href="http://yeticave/lot.php?id=<?=$index?>"><?=strip_tags($lot["name"]);?></a></h3>
+                        <h3 class="lot__title"><a class="text-link" href="lot.php?id=<?=$index?>"><?=strip_tags($lot["name"]);?></a></h3>
 
                         <div class="lot__state">
                             <div class="lot__rate">
